chatcal-ai/pyproject.toml
[tool.poetry]
name = "chatcal-ai"
version = "0.1.0"
description = "A friendly AI chatbot for booking Google Calendar appointments"
authors = ["pgits.job@gmail.com"]
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.11.0"
llama-index-llms-anthropic = "^0.3.0"
llama-index-llms-gemini = "^0.3.0"
google-generativeai = "^0.8.0"
llama-index-tools-google = "^0.2.0"
google-api-python-client = "^2.100.0"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
redis = "^5.0.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.10.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true